Moves court end time to the previous screent
