+++ apps/judicial-system/web/src/routes/Judge/CourtRecord/CourtRecord.spec.tsx
@@ -27,41 +27,37 @@ describe('/domari-krafa/thingbok', () => {
         mocks={[
           ...mockCaseQueries,
           ...mockJudgeQuery,
-          ...mockUpdateCaseMutation([
-            {
-              id: 'test_id_9',
-              courtStartDate: '2020-09-16T19:51:28.000Z',
-            } as UpdateCase,
-            {
-              id: 'test_id_9',
-              courtAttendees:
-                'Ruth Bader Ginsburg saksóknari\nJon Harring kærði\nSaul Goodman skipaður verjandi kærða',
-            } as UpdateCase,
-            {
-              id: 'test_id_9',
-              policeDemands:
-                'Þess er krafist að Jon Harring, kt. 000000-0000, sæti gæsluvarðhaldi með úrskurði string, til miðvikudagsins 16. september 2020, kl. 00:00.',
-            } as UpdateCase,
-            {
-              id: 'test_id_9',
-              policeDemands:
-                'Þess er krafist að Jon Harring, kt. 111111-1110, sæti gæsluvarðhaldi með úrskurði string, til miðvikudagsins 16. september 2020, kl. 00:00.',
-            } as UpdateCase,
-            {
-              id: 'test_id_9',
-              accusedPleaDecision: AccusedPleaDecision.ACCEPT,
-            } as UpdateCase,
-            {
-              id: 'test_id_9',
-              accusedPleaAnnouncement:
-                'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Iam id ipsum absurdum, maximum malum neglegi. Sed ne, dum huic obsequor, vobis molestus sim. Quae dici eadem de ceteris virtutibus possunt, quarum omnium fundamenta vos in voluptate tamquam in aqua ponitis. Hanc ergo intuens debet institutum illud quasi signum absolvere. Duo Reges: constructio interrete. Quorum sine causa fieri nihil putandum est. Antiquorum autem sententiam Antiochus noster mihi videtur persequi diligentissime, quam eandem Aristoteli fuisse et Polemonis docet. Atque ab his initiis profecti omnium virtutum et originem et progressionem persecuti sunt. Nam et complectitur verbis, quod vult, et dicit plane, quod intellegam; Cur deinde Metrodori liberos commendas?',
-            } as UpdateCase,
-            {
-              id: 'test_id_9',
-              litigationPresentations:
-                'Sækjandi ítrekar kröfu um gæsluvarðhald, reifar og rökstyður kröfuna og leggur málið í úrskurð með venjulegum fyrirvara.\n\nVerjandi kærða ítrekar mótmæli hans, krefst þess að kröfunni verði hafnað, til vara að kærða verði gert að sæta farbanni í stað gæsluvarðhalds, en til þrautavara að gæsluvarðhaldi verði markaður skemmri tími en krafist er og að kærða verði ekki gert að sæta einangrun á meðan á gæsluvarðhaldi stendur. Verjandinn reifar og rökstyður mótmælin og leggur málið í úrskurð með venjulegum fyrirvara.',
-            } as UpdateCase,
-          ]),
+          ...mockUpdateCaseMutation(
+            [
+              {
+                courtStartDate: '2020-09-16T19:51:28.000Z',
+              } as UpdateCase,
+              {
+                courtAttendees:
+                  'Ruth Bader Ginsburg saksóknari\nJon Harring kærði\nSaul Goodman skipaður verjandi kærða',
+              } as UpdateCase,
+              {
+                policeDemands:
+                  'Þess er krafist að Jon Harring, kt. 000000-0000, sæti gæsluvarðhaldi með úrskurði string, til miðvikudagsins 16. september 2020, kl. 00:00.',
+              } as UpdateCase,
+              {
+                policeDemands:
+                  'Þess er krafist að Jon Harring, kt. 111111-1110, sæti gæsluvarðhaldi með úrskurði string, til miðvikudagsins 16. september 2020, kl. 00:00.',
+              } as UpdateCase,
+              {
+                accusedPleaDecision: AccusedPleaDecision.ACCEPT,
+              } as UpdateCase,
+              {
+                accusedPleaAnnouncement:
+                  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Iam id ipsum absurdum, maximum malum neglegi. Sed ne, dum huic obsequor, vobis molestus sim. Quae dici eadem de ceteris virtutibus possunt, quarum omnium fundamenta vos in voluptate tamquam in aqua ponitis. Hanc ergo intuens debet institutum illud quasi signum absolvere. Duo Reges: constructio interrete. Quorum sine causa fieri nihil putandum est. Antiquorum autem sententiam Antiochus noster mihi videtur persequi diligentissime, quam eandem Aristoteli fuisse et Polemonis docet. Atque ab his initiis profecti omnium virtutum et originem et progressionem persecuti sunt. Nam et complectitur verbis, quod vult, et dicit plane, quod intellegam; Cur deinde Metrodori liberos commendas?',
+              } as UpdateCase,
+              {
+                litigationPresentations:
+                  'Sækjandi ítrekar kröfu um gæsluvarðhald, reifar og rökstyður kröfuna og leggur málið í úrskurð með venjulegum fyrirvara.\n\nVerjandi kærða ítrekar mótmæli hans, krefst þess að kröfunni verði hafnað, til vara að kærða verði gert að sæta farbanni í stað gæsluvarðhalds, en til þrautavara að gæsluvarðhaldi verði markaður skemmri tími en krafist er og að kærða verði ekki gert að sæta einangrun á meðan á gæsluvarðhaldi stendur. Verjandinn reifar og rökstyður mótmælin og leggur málið í úrskurð með venjulegum fyrirvara.',
+              } as UpdateCase,
+            ],
+            'test_id_9',
+          ),
         ]}
         addTypename={false}
       >
@@ -116,22 +112,22 @@ describe('/domari-krafa/thingbok', () => {
         mocks={[
           ...mockCaseQueries,
           ...mockJudgeQuery,
-          ...mockUpdateCaseMutation([
-            {
-              id: 'test_id_10',
-              courtStartDate: '2020-09-16T19:51:28.000Z',
-            } as UpdateCase,
-            {
-              id: 'test_id_10',
-              courtAttendees:
-                'Ruth Bader Ginsburg saksóknari\nJon Harring kærði\nSaul Goodman skipaður verjandi kærða',
-            } as UpdateCase,
-            {
-              id: 'test_id_10',
-              policeDemands:
-                'Þess er krafist að Jon Harring, kt. string, sæti farbanni með úrskurði string, til miðvikudagsins 16. september 2020, kl. 00:00.',
-            } as UpdateCase,
-          ]),
+          ...mockUpdateCaseMutation(
+            [
+              {
+                courtStartDate: '2020-09-16T19:51:28.000Z',
+              } as UpdateCase,
+              {
+                courtAttendees:
+                  'Ruth Bader Ginsburg saksóknari\nJon Harring kærði\nSaul Goodman skipaður verjandi kærða',
+              } as UpdateCase,
+              {
+                policeDemands:
+                  'Þess er krafist að Jon Harring, kt. string, sæti farbanni með úrskurði string, til miðvikudagsins 16. september 2020, kl. 00:00.',
+              } as UpdateCase,
+            ],
+            'test_id_10',
+          ),
         ]}
         addTypename={false}
       >
