+++ apps/judicial-system/web/src/routes/Judge/Ruling/StepOne/RulingStepOne.spec.tsx
@@ -30,25 +30,24 @@ describe('/domari-krafa/urskurdur', () => {
         mocks={[
           ...mockCaseQueries,
           ...mockJudgeQuery,
-          ...mockUpdateCaseMutation([
-            {
-              id: 'test_id_3',
-              ruling:
-                'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Non igitur bene. Idem fecisset Epicurus, si sententiam hanc, quae nunc Hieronymi est, coniunxisset cum Aristippi vetere sententia. Respondent extrema primis, media utrisque, omnia omnibus. Nam prius a se poterit quisque discedere quam appetitum earum rerum, quae sibi conducant, amittere. Duo Reges: constructio interrete. Sed quae tandem ista ratio est?',
-            } as UpdateCase,
-            {
-              id: 'test_id_3',
-              custodyRestrictions: [CaseCustodyRestrictions.MEDIA],
-            } as UpdateCase,
-            {
-              id: 'test_id_3',
-              custodyEndDate: '2020-10-24T12:31:00Z',
-            } as UpdateCase,
-            {
-              id: 'test_id_3',
-              decision: CaseDecision.ACCEPTING,
-            } as UpdateCase,
-          ]),
+          ...mockUpdateCaseMutation(
+            [
+              {
+                ruling:
+                  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Non igitur bene. Idem fecisset Epicurus, si sententiam hanc, quae nunc Hieronymi est, coniunxisset cum Aristippi vetere sententia. Respondent extrema primis, media utrisque, omnia omnibus. Nam prius a se poterit quisque discedere quam appetitum earum rerum, quae sibi conducant, amittere. Duo Reges: constructio interrete. Sed quae tandem ista ratio est?',
+              } as UpdateCase,
+              {
+                custodyRestrictions: [CaseCustodyRestrictions.MEDIA],
+              } as UpdateCase,
+              {
+                custodyEndDate: '2020-10-24T12:31:00Z',
+              } as UpdateCase,
+              {
+                decision: CaseDecision.ACCEPTING,
+              } as UpdateCase,
+            ],
+            'test_id_3',
+          ),
         ]}
         addTypename={false}
       >
@@ -99,15 +98,17 @@ describe('/domari-krafa/urskurdur', () => {
         mocks={[
           ...mockCaseQueries,
           ...mockJudgeQuery,
-          ...mockUpdateCaseMutation([
-            {
-              id: 'test_id',
-              custodyRestrictions: [
-                CaseCustodyRestrictions.ISOLATION,
-                CaseCustodyRestrictions.MEDIA,
-              ],
-            } as UpdateCase,
-          ]),
+          ...mockUpdateCaseMutation(
+            [
+              {
+                custodyRestrictions: [
+                  CaseCustodyRestrictions.ISOLATION,
+                  CaseCustodyRestrictions.MEDIA,
+                ],
+              } as UpdateCase,
+            ],
+            'test_id',
+          ),
         ]}
         addTypename={false}
       >
@@ -138,12 +139,14 @@ describe('/domari-krafa/urskurdur', () => {
         mocks={[
           ...mockCaseQueries,
           ...mockJudgeQuery,
-          ...mockUpdateCaseMutation([
-            {
-              id: 'test_id',
-              decision: CaseDecision.REJECTING,
-            } as UpdateCase,
-          ]),
+          ...mockUpdateCaseMutation(
+            [
+              {
+                decision: CaseDecision.REJECTING,
+              } as UpdateCase,
+            ],
+            'test_id',
+          ),
         ]}
         addTypename={false}
       >
@@ -181,12 +184,14 @@ describe('/domari-krafa/urskurdur', () => {
         mocks={[
           ...mockCaseQueries,
           ...mockJudgeQuery,
-          ...mockUpdateCaseMutation([
-            {
-              id: 'test_id',
-              decision: CaseDecision.ACCEPTING_ALTERNATIVE_TRAVEL_BAN,
-            } as UpdateCase,
-          ]),
+          ...mockUpdateCaseMutation(
+            [
+              {
+                decision: CaseDecision.ACCEPTING_ALTERNATIVE_TRAVEL_BAN,
+              } as UpdateCase,
+            ],
+            'test_id',
+          ),
         ]}
         addTypename={false}
       >
