+++ apps/judicial-system/web/src/routes/Judge/Ruling/StepTwo/RulingStepTwo.treat.ts
@@ -1,0 +1,8 @@
+import { theme } from '@island.is/island-ui/theme'
+import { style } from 'treat'
+
+export const courtEndTimeContainer = style({
+  marginBottom: theme.spacing[5],
+  paddingBottom: theme.spacing[5],
+  borderBottom: `2px solid ${theme.color.purple100}`,
+})
