+++ apps/judicial-system/web/src/routes/Judge/Ruling/StepTwo/RulingStepTwo.tsx
@@ -1,6 +1,7 @@
 import {
   Box,
   GridColumn,
+  GridContainer,
   GridRow,
   Input,
   RadioButton,
@@ -13,6 +14,7 @@ import {
   BlueBox,
   CaseNumbers,
   FormContentContainer,
+  TimeInputField,
 } from '@island.is/judicial-system-web/src/shared-components'
 import {
   Case,
@@ -33,6 +35,9 @@ import {
   validateAndSendToServer,
   removeTabsValidateAndSet,
   setCheckboxAndSendToServer,
+  validateAndSetTime,
+  validateAndSendTimeToServer,
+  getTimeFromDate,
 } from '@island.is/judicial-system-web/src/utils/formHelper'
 import CheckboxList from '@island.is/judicial-system-web/src/shared-components/CheckboxList/CheckboxList'
 import {
@@ -42,16 +47,24 @@ import {
 import {
   capitalize,
   formatAccusedByGender,
+  formatDate,
   NounCases,
+  TIME_FORMAT,
 } from '@island.is/judicial-system/formatters'
 import { getConclusion } from '@island.is/judicial-system-web/src/utils/stepHelper'
 import { useRouter } from 'next/router'
 import useCase from '@island.is/judicial-system-web/src/utils/hooks/useCase'
+import useDateTime from '@island.is/judicial-system-web/src/utils/hooks/useDateTime'
+import * as style from './RulingStepTwo.treat'
 
 export const RulingStepTwo: React.FC = () => {
   const router = useRouter()
   const id = router.query.id
   const [workingCase, setWorkingCase] = useState<Case>()
+  const [
+    courtDocumentEndErrorMessage,
+    setCourtDocumentEndErrorMessage,
+  ] = useState<string>('')
 
   const { updateCase } = useCase()
   const { data, loading } = useQuery(CaseQuery, {
@@ -59,6 +72,10 @@ export const RulingStepTwo: React.FC = () => {
     fetchPolicy: 'no-cache',
   })
   const resCase = data?.case
+
+  const { isValidTime: isValidCourtEndTime } = useDateTime({
+    time: getTimeFromDate(workingCase?.courtEndTime),
+  })
 
   useEffect(() => {
     document.title = 'Úrskurðarorð - Réttarvörslugátt'
@@ -534,6 +551,58 @@ export const RulingStepTwo: React.FC = () => {
                 </Text>
               )}
             </Box>
+            <Box className={style.courtEndTimeContainer}>
+              <Box marginBottom={2}>
+                <Text as="h3" variant="h3">
+                  Þinghald
+                </Text>
+              </Box>
+              <GridContainer>
+                <GridRow>
+                  <GridColumn>
+                    <TimeInputField
+                      onChange={(evt) =>
+                        validateAndSetTime(
+                          'courtEndTime',
+                          workingCase.courtStartDate,
+                          evt.target.value,
+                          ['empty', 'time-format'],
+                          workingCase,
+                          setWorkingCase,
+                          courtDocumentEndErrorMessage,
+                          setCourtDocumentEndErrorMessage,
+                        )
+                      }
+                      onBlur={(evt) =>
+                        validateAndSendTimeToServer(
+                          'courtEndTime',
+                          workingCase.courtStartDate,
+                          evt.target.value,
+                          ['empty', 'time-format'],
+                          workingCase,
+                          updateCase,
+                          setCourtDocumentEndErrorMessage,
+                        )
+                      }
+                    >
+                      <Input
+                        data-testid="courtEndTime"
+                        name="courtEndTime"
+                        label="Þinghaldi lauk (kk:mm)"
+                        placeholder="Veldu tíma"
+                        defaultValue={formatDate(
+                          workingCase.courtEndTime,
+                          TIME_FORMAT,
+                        )}
+                        errorMessage={courtDocumentEndErrorMessage}
+                        hasError={courtDocumentEndErrorMessage !== ''}
+                        required
+                      />
+                    </TimeInputField>
+                  </GridColumn>
+                </GridRow>
+              </GridContainer>
+            </Box>
           </FormContentContainer>
           <FormContentContainer isFooter>
             <FormFooter
@@ -541,7 +610,8 @@ export const RulingStepTwo: React.FC = () => {
               nextUrl={`${Constants.CONFIRMATION_ROUTE}/${id}`}
               nextIsDisabled={
                 !workingCase.accusedAppealDecision ||
-                !workingCase.prosecutorAppealDecision
+                !workingCase.prosecutorAppealDecision ||
+                !isValidCourtEndTime?.isValid
               }
             />
           </FormContentContainer>
