+++ apps/judicial-system/web/src/routes/Prosecutor/CreateDetentionRequest/StepTwo/StepTwo.spec.tsx
@@ -37,25 +37,25 @@ describe('Create detention request, step two', () => {
           ...mockCaseQueries,
           ...mockProsecutorQuery,
           ...mockUsersQuery,
-          ...mockUpdateCaseMutation([
-            {
-              arrestDate: '2020-11-15',
-            } as UpdateCase,
-            {
-              id: 'test_id_6',
-              arrestDate: '2020-11-15T13:37:00Z',
-            } as UpdateCase,
-            {
-              id: 'test_id_6',
-              requestedCourtDate: formatISO(lastDateOfTheMonth, {
-                representation: 'date',
-              }),
-            } as UpdateCase,
-            {
-              id: 'test_id_6',
-              requestedCourtDate: formatISO(lastDateOfTheMonth),
-            } as UpdateCase,
-          ]),
+          ...mockUpdateCaseMutation(
+            [
+              {
+                arrestDate: '2020-11-15',
+              } as UpdateCase,
+              {
+                arrestDate: '2020-11-15T13:37:00Z',
+              } as UpdateCase,
+              {
+                requestedCourtDate: formatISO(lastDateOfTheMonth, {
+                  representation: 'date',
+                }),
+              } as UpdateCase,
+              {
+                requestedCourtDate: formatISO(lastDateOfTheMonth),
+              } as UpdateCase,
+            ],
+            'test_id_6',
+          ),
         ]}
         addTypename={false}
       >
@@ -186,10 +186,19 @@ describe('Create detention request, step two', () => {
 
     const tomorrow = addDays(new Date(), 1).getDate().toString()
 
+    const candidateTomorrows = arrestWrapper.getAllByText(tomorrow)
+
+    expect(candidateTomorrows.length).toBeGreaterThan(0)
+
+    // Make sure we pick a date from the curent month
+    let tommorrowOfCurentMont: HTMLElement
+    if (tomorrow > '16') {
+      tommorrowOfCurentMont = candidateTomorrows[candidateTomorrows.length - 1]
+    } else {
+      tommorrowOfCurentMont = candidateTomorrows[0]
+    }
+
     // Assert
-    expect(arrestWrapper.getAllByText(tomorrow)[0]).toHaveAttribute(
-      'aria-disabled',
-      'true',
-    )
+    expect(tommorrowOfCurentMont).toHaveAttribute('aria-disabled', 'true')
   })
 })
